User Interface/requirements.txt
absl-py==0.9.0
astor==0.8.1
audioread==2.1.8
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
decorator==4.4.1
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.1
h5py==2.10.0
idna==2.8
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
librosa==0.7.2
llvmlite==0.31.0
Markdown==3.2
numba==0.48.0
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
pip==20.0.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pyrtlsdr==0.2.91
requests==2.22.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
setuptools==45.2.0
six==1.14.0
SoundFile==0.10.3.post1
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
urllib3==1.25.8
Werkzeug==1.0.0
wheel==0.34.2
wrapt==1.11.2
